+++ Sales Management App/Sales Management App/Presentation/Tabs/Aggregation/AggregationSnapshot.cs
@@ -1,15 +1,17 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using SalesManagementApp.Core.Application.Models;
 
-namespace Sales_Management_App
-{
-    internal class AggregationSnapshot
-    {
+namespace Sales_Management_App.Presentation.Tabs.Aggregation {
+    internal sealed class AggregationSnapshot {
         public DateTime StartDate { get; set; }
+
         public DateTime EndDate { get; set; }
+
         public List<ProductSalesSummary> ProductSummaries { get; set; }
+
         public List<WeeklySalesSummary> WeeklySummaries { get; set; }
+
         public int TotalSalesAmount { get; set; }
     }
 }
